Add t/RT74518.t, checking RT#74518
RT#74518 is a bug reported against AnyEvent::MessagePack (in the
AnyEvent-MPRPC RT queue). I was suspecting that the issue may be in
Data::MessagePack::Stream. But as the this tests shows (as fas as it
doesn't have a bug itself), this is not the case as it succeeds with
the current code.
Anyway, that's one more testcase that may avoid regressions later (or
be useful to write other testcases), so let's add it to the testsuite.